Small improvements

+++ KeysPerSecond/src/me/roan/kps/Menu.java
@@ -590,7 +590,7 @@ public class Menu{
 		TotPanel.hits = 0;
 		Main.reconfigure();
 		Main.mainLoop();
-		KeyInformation.autoIndex = (int)(Main.config.keyinfo.size() * 2 - 2);
+		KeyInformation.autoIndex = Main.config.keyinfo.size() * 2 - 2;
 		Main.graphFrame.setAlwaysOnTop(Main.config.overlay);
 		Main.frame.setAlwaysOnTop(Main.config.overlay);
 	}
